% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{billings_cdm}
\alias{billings_cdm}
\title{CDM item status of patients in appointment view}
\usage{
billings_cdm(
  dMeasureCDM_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  intID = NULL,
  intID_Date = Sys.Date(),
  cdm_chosen = cdm_item_names,
  itemstatus_chosen = unlist(item_status, use.names = FALSE),
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasureCDM_obj}{dMeasureCDM R6 object}

\item{date_from}{$date_a (from dMeasure) start date}

\item{date_to}{$date_b end date (inclusive)}

\item{clinicians}{$clinicians (from dMeasure) list of clinicians to view}

\item{intID}{vector of internal ID (default is NULL, in which case appointments_list is used)}

\item{intID_Date}{if intID is not NULL, then date to check (default is Sys.Date())
intID_Date can be a single value (applies to all intID), or a vector (different
date can be defined for each intID) the same length as intID.}

\item{cdm_chosen}{(defaut cdm_item_names) item types to show, defaults to all available}

\item{itemstatus_chosen}{(default is contents of item_status) item status to show, defaults to all available}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
  returns empty dataframe if no valid subscription for chosen clinicians
  and selected date-range
}
\description{
filter to CDM item billed prior to (or on) the day of displayed appointments
only show most recent billed item in each category
}
