% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Chronic disease management list module - server}
\usage{
datatableServer(id, dMCDM)
}
\arguments{
\item{id}{id}

\item{dMCDM}{dMeasureCDM R6 object
access to appointments lists, condition lists, history and EMR database}
}
\value{
none
}
\description{
chronic disease management items claimed, pending or unclaimed
for appointment list
}
