% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDM.R
\name{dMeasureCDM}
\alias{dMeasureCDM}
\title{dMeasureCDM class}
\description{
list chronic disease management opportunities
}
\details{
dMeasureCDM class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dMeasureCDM$new()}}
\item \href{#method-aha75_list_cdm}{\code{dMeasureCDM$aha75_list_cdm()}}
\item \href{#method-diabetes_list_cdm}{\code{dMeasureCDM$diabetes_list_cdm()}}
\item \href{#method-asthma_list_cdm}{\code{dMeasureCDM$asthma_list_cdm()}}
\item \href{#method-malignancy_list_cdm}{\code{dMeasureCDM$malignancy_list_cdm()}}
\item \href{#method-hiv_list_cdm}{\code{dMeasureCDM$hiv_list_cdm()}}
\item \href{#method-haemoglobinopathy_list_cdm}{\code{dMeasureCDM$haemoglobinopathy_list_cdm()}}
\item \href{#method-asplenic_list_cdm}{\code{dMeasureCDM$asplenic_list_cdm()}}
\item \href{#method-transplant_list_cdm}{\code{dMeasureCDM$transplant_list_cdm()}}
\item \href{#method-trisomy21_list_cdm}{\code{dMeasureCDM$trisomy21_list_cdm()}}
\item \href{#method-chroniclungdisease_list_cdm}{\code{dMeasureCDM$chroniclungdisease_list_cdm()}}
\item \href{#method-neurologic_list_cdm}{\code{dMeasureCDM$neurologic_list_cdm()}}
\item \href{#method-chronicliverdisease_list_cdm}{\code{dMeasureCDM$chronicliverdisease_list_cdm()}}
\item \href{#method-chronicrenaldisease_list_cdm}{\code{dMeasureCDM$chronicrenaldisease_list_cdm()}}
\item \href{#method-cardiacdisease_list_cdm}{\code{dMeasureCDM$cardiacdisease_list_cdm()}}
\item \href{#method-billings_cdm}{\code{dMeasureCDM$billings_cdm()}}
\item \href{#method-clone}{\code{dMeasureCDM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$new(dMeasure_obj, dMeasureBillings_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aha75_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-aha75_list_cdm}{}}}
\subsection{Method \code{aha75_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$aha75_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diabetes_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-diabetes_list_cdm}{}}}
\subsection{Method \code{diabetes_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$diabetes_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asthma_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-asthma_list_cdm}{}}}
\subsection{Method \code{asthma_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$asthma_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-malignancy_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-malignancy_list_cdm}{}}}
\subsection{Method \code{malignancy_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$malignancy_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hiv_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-hiv_list_cdm}{}}}
\subsection{Method \code{hiv_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$hiv_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-haemoglobinopathy_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-haemoglobinopathy_list_cdm}{}}}
\subsection{Method \code{haemoglobinopathy_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$haemoglobinopathy_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asplenic_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-asplenic_list_cdm}{}}}
\subsection{Method \code{asplenic_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$asplenic_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transplant_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-transplant_list_cdm}{}}}
\subsection{Method \code{transplant_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$transplant_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trisomy21_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-trisomy21_list_cdm}{}}}
\subsection{Method \code{trisomy21_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$trisomy21_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroniclungdisease_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-chroniclungdisease_list_cdm}{}}}
\subsection{Method \code{chroniclungdisease_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$chroniclungdisease_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neurologic_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-neurologic_list_cdm}{}}}
\subsection{Method \code{neurologic_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$neurologic_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chronicliverdisease_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-chronicliverdisease_list_cdm}{}}}
\subsection{Method \code{chronicliverdisease_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$chronicliverdisease_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chronicrenaldisease_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-chronicrenaldisease_list_cdm}{}}}
\subsection{Method \code{chronicrenaldisease_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$chronicrenaldisease_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cardiacdisease_list_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-cardiacdisease_list_cdm}{}}}
\subsection{Method \code{cardiacdisease_list_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$cardiacdisease_list_cdm(intID_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-billings_cdm"></a>}}
\if{latex}{\out{\hypertarget{method-billings_cdm}{}}}
\subsection{Method \code{billings_cdm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$billings_cdm(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  intID = NULL,
  intID_Date = Sys.Date(),
  cdm_chosen = cdm_item_names,
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureCDM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
