% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_introduction_df}
\alias{steps_introduction_df}
\title{steps_introduction_df}
\usage{
steps_introduction_df(element_name = as.character(NA))
}
\arguments{
\item{element_name}{the HTML datatable element, including datatable and helpers such as print/copy view}
}
\description{
returns a dataframe of rintrojs steps

requires pro-forma steps from DailyMeasure
}
\details{
rintrojs steps for CDM module
}
