% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDM.R
\docType{data}
\name{dMeasureCDM}
\alias{dMeasureCDM}
\title{dMeasureCDM class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dMeasureCDM
}
\description{
list chronic disease management opportunities
}
\details{
dMeasureCDM class
}
\keyword{datasets}
